pyproject.toml
[project]
name = "tech_europe_hackathon"
version = "0.1.0"
description = "Simplified AI text generation and modification system"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=4.0.0",
    "crewai[tools]>=0.22.0",
    "elevenlabs>=0.2.26",
    "python-dotenv>=1.0.0",
    "openai>=1.3.0",
    "aci-sdk>=0.1.0",
    "weaviate-client>=4.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0"
]

[project.scripts]
tech-europe-hackathon = "app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tech_europe_hackathon"]
